Added bifrost links and moved structure
