+++ jobmon_gui/local_testing/jobmon_gui/testing_servers/functionaltest_server.py
@@ -28,7 +28,7 @@ class WebServerProcess:
         os.environ["JOBMON__OTLP__WEB_ENABLED"] = "true"
         os.environ["JOBMON__OTLP__SPAN_EXPORTER"] = ""
         os.environ["JOBMON__OTLP__LOG_EXPORTER"] = ""
-        os.environ["JOBMON__HTTP__SERVICE_URL"] = "http://localhost:8070"
+        os.environ["JOBMON__HTTP__SERVICE_URL"] = "http://localhost:8070/api/v2"
 
         if not os.path.exists(self.filepath):
             open(self.filepath, 'a').close()
@@ -37,7 +37,7 @@ class WebServerProcess:
 
         config = JobmonConfig(dict_config={"db": {"sqlalchemy_database_uri": database_uri}})
         app = get_app(config)
-        config.set("http", "service_url", f"http://{self.web_host}:{self.web_port}")
+        config.set("http", "service_url", f"http://{self.web_host}:{self.web_port}/api/v2")
         config.write()
 
         with app.app_context():
