+++ jobmon_gui/src/assets/logo/ihme_icon.png
