+++ jobmon_gui/src/assets/logo/ihme_logo_white_text.png
