+++ jobmon_gui/src/assets/logo/logo.ts
