+++ jobmon_gui/src/components/JobmonProgressBar.tsx
@@ -3,7 +3,7 @@ import ProgressBar from 'react-bootstrap/ProgressBar';
 import { OverlayTrigger } from "react-bootstrap";
 import Popover from 'react-bootstrap/Popover';
 
-import '../css/jobmon_gui.css';
+import '../styles/jobmon_gui.css';
 
 export default function JobmonProgressBar({tasks, pending, scheduled, running, done, fatal, num_attempts_avg, num_attempts_min, num_attempts_max, maxc, placement, style="striped"}) {
     num_attempts_avg = parseFloat(num_attempts_avg).toFixed(1);
