+++ jobmon_gui/src/components/navigation/PageNavigation.tsx
@@ -18,7 +18,7 @@ import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import {PropsWithChildren, useState} from "react";
 import {Link, useNavigate} from "react-router-dom";
-import {IhmeIcon} from "./logo/logo";
+import {IhmeIcon} from "../../assets/logo/logo";
 
 import FeedIcon from "@mui/icons-material/Feed";
 import Typography from "@mui/material/Typography";
@@ -27,7 +27,6 @@ import MediationIcon from '@mui/icons-material/Mediation';
 import BifrostLinks from "./BifrostLinks";
 import BatchPredictionIcon from '@mui/icons-material/BatchPrediction';
 import {Tooltip} from "@mui/material";
-import HelpCenterIcon from '@mui/icons-material/HelpCenter';
 import bifrostEnabled from "./Bifrost";
 import "@fontsource/archivo"
