+++ jobmon_gui/src/screens/task_details_page/node_list.tsx
