+++ jobmon_gui/src/screens/workflow_details_page/memory_histogram.tsx
