+++ jobmon_gui/src/screens/workflow_details_page/runtime_histogram.tsx
