+++ jobmon_gui/src/screens/workflow_details_page/tasks.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import TaskTable from './task_table';
 import { FaSearch } from "react-icons/fa";
 
-import { safe_rum_start_span, safe_rum_unit_end } from '../../utilities/rum'
+import { safe_rum_start_span, safe_rum_unit_end } from '../../utils/rum'
 
 export default function Tasks({ tasks, onSubmit, register, loading, apm }) {
     useEffect(() => {
