+++ jobmon_gui/src/screens/workflow_overview_page/wf_table.tsx
@@ -8,7 +8,7 @@ import ToolkitProvider, { CSVExport } from 'react-bootstrap-table2-toolkit/dist/
 import "react-bootstrap-table-next/dist/react-bootstrap-table2.min.css"
 import Spinner from 'react-bootstrap/Spinner';
 import { Link } from "react-router-dom";
-import { convertDate, convertDatePST } from '../../utilities/formatters';
+import { convertDate, convertDatePST } from '../../utils/formatters';
 import { FaCaretDown, FaCaretUp, FaCircle } from "react-icons/fa";
 
 // @ts-ignore
